refactor app.py to launch both http and ws servers
