+++ backend/app.py
@@ -1,11 +1,12 @@
-from flask import Flask, request, jsonify
+from flask import Flask
 from flask_cors import CORS
 import os
-import uuid
-import random
-import string
-from face_extractor.face_extractor import extract_face
+import asyncio
+from threading import Thread
+
+from numpy import False_
 from lobby.lobby_controller import lobby_bp
+from lobby.lobby_websockets import socket_serve
 
 app = Flask(__name__)
 CORS(app)  # Enable CORS for cross-origin requests
@@ -22,26 +23,17 @@ app.register_blueprint(lobby_bp)
 os.makedirs(RAW_FOLDER, exist_ok=True)
 os.makedirs(PROCESSED_FOLDER, exist_ok=True)
 
-@app.route('/upload', methods=['POST'])
-def upload_file():
-    if 'file' not in request.files:
-        return jsonify({"error": "No file part"}), 400
+async def start_flask():
+    print("Starting the flask server")
+    app.run(debug=False, host="0.0.0.0")
+    return 1
 
-    file = request.files['file']
-
-    if file.filename == '':
-        return jsonify({"error": "No selected file"}), 400
-
-    if file:
-        # Store file in raw directory
-        file_path = os.path.join(app.config['RAW_FOLDER'], file.filename)
-        file.save(file_path)
-
-        # Extract face and store in processed directory
-        output_path = os.path.join(app.config['PROCESSED_FOLDER'], file.filename)
-        extract_face(file_path, output_path)
-
-        return jsonify({"message": "File uploaded successfully", "file_path": file_path}), 200
+async def start_servers():
+    tasks = list()
+    if not (app.debug or os.environ.get('FLASK_ENV') == 'development') or os.environ.get('WERKZEUG_RUN_MAIN') == 'true':
+        tasks.append(asyncio.create_task(socket_serve()))
+    tasks.append(asyncio.create_task(start_flask()))
+    await asyncio.gather(*tasks)
 
 if __name__ == '__main__':
-    app.run(debug=True, host="0.0.0.0")
+    asyncio.run(start_servers())
